Crate sui_transaction_builder

Modules

unresolved

Structs

Function

A separate type to support denoting a function by a more structured representation.

Serialized

A transaction input that will be serialized from BCS.

TransactionBuilder

A builder for creating transactions. Use resolve to finalize the transaction data.