sui_transaction_builder
Module unresolved

Structs

GasPayment

Input

A potentially unresolved transaction input. Note that one can construct a fully resolved input using the provided constructors, but this struct is also useful when the input data is not complete.

ObjectReference

ProgrammableTransaction

Transaction

Enums

InputKind

Value